restructuring code to improve modularity
