--- backend/experiment/start/a.js
@@ -1,1 +1,0 @@
-console.log("a");
