--- backend/experiment/start/b/b.js
@@ -1,1 +1,0 @@
-console.log("b");
