--- backend/experiment/start/b/c/c.js
@@ -1,1 +1,0 @@
-console.log("C");
