+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
-import './App.css'
+import './styles/app.css'
 
 function App() {
   const [count, setCount] = useState(0)
