+++ frontend/src/components/user/Profile.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import axios from "axios";
-import "./profile.css";
-import Navbar from "../Navbar";
+import "../../styles/pages/profile.css";
+import Navbar from "../layout/Navbar";
 import { UnderlineNav } from "@primer/react";
 import { BookIcon, RepoIcon } from "@primer/octicons-react";
 import HeatMapProfile from "./HeatMap";
