+++ frontend/src/components/utils/Hr.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import "./Hr.css"
+import "../../styles/components/hr.css"
 export default function Hr(){
     return(
         <hr id="hr"/>
