+++ frontend/src/components/utils/RepoTable.jsx
@@ -1,7 +1,7 @@
 import React,{useState} from "react";
 import Avatar from '../user/UserAvatar';
-import "./RepoTable.css"
-import "../../assets/colors.css";
+import "../../styles/components/repoTable.css"
+// import "../../assets/colors.css";
 import { Link } from "react-router-dom";
 
 export default function RepositoryTable({rows, reponame, username}) {
