+++ frontend/src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './main.css'
-import "./assets/colors.css"
+import './styles/main.css'
+import "./styles/base/colors.css"
 import App from './App.jsx'
 import { AuthProvider } from './authContext.jsx'
 import ProjectRoutes from './Routes.jsx'
