+++ frontend/src/styles/components/hr.css
