+++ frontend/src/styles/pages/profile.css
